Case Book Servlet Add New

+++ Module_3/#Case/LibraryManagement/src/main/java/com/librarymanagement/components/user/services/IUserDAO.java
@@ -3,7 +3,7 @@ package com.librarymanagement.components.user.services;
 import com.librarymanagement.components.user.models.User;
 import com.librarymanagement.services.IAbstractService;
 
-public interface IUserDAO extends IAbstractService<User,Integer> {
+public interface IUserDAO extends IAbstractService<User,Long> {
 
     boolean add(User newUser);
